Rename module material to materials
